refactor(sensors): merge baro_thrust_estimator into VehicleAirData
Move online thrust compensation estimation (CF+RLS) from standalone
module into VehicleAirData, eliminating the separate process, init
script entries, and uORB hop.

- baro_thrust_cf_rls math library stays cleanly separated with its
  own unit test
- Estimation runs inline after baro publish when SENS_BAR_AUTOCAL
  bit 1 is set
- Parameter save on disarm, convergence logic unchanged
- Remove baro_thrust_estimator module, Kconfig, init script blocks
