+++ ROMFS/px4fmu_common/init.d/rc.mc_apps
@@ -30,12 +30,6 @@ fi
 #
 mc_hover_thrust_estimator start
 
-# Bit 1 of SENS_BAR_AUTOCAL enables online thrust compensation (value > 1)
-if param greater -s SENS_BAR_AUTOCAL 1
-then
-	baro_thrust_estimator start
-fi
-
 flight_mode_manager start
 mc_pos_control start
