+++ ROMFS/px4fmu_common/init.d/rc.vtol_apps
@@ -20,12 +20,6 @@ flight_mode_manager start
 mc_pos_control start vtol
 mc_hover_thrust_estimator start
 
-# Bit 1 of SENS_BAR_AUTOCAL enables online thrust compensation (value > 1)
-if param greater -s SENS_BAR_AUTOCAL 1
-then
-	baro_thrust_estimator start
-fi
-
 if param greater -s MC_AT_EN 0
 then
 	mc_autotune_attitude_control start
