--- src/modules/baro_thrust_estimator/Kconfig
@@ -1,12 +1,0 @@
-menuconfig MODULES_BARO_THRUST_ESTIMATOR
-	bool "baro_thrust_estimator"
-	default n
-	---help---
-		Enable support for baro_thrust_estimator
-
-menuconfig USER_BARO_THRUST_ESTIMATOR
-	bool "baro_thrust_estimator running as userspace module"
-	default y
-	depends on BOARD_PROTECTED && MODULES_BARO_THRUST_ESTIMATOR
-	---help---
-		Put baro_thrust_estimator in userspace memory
