+++ src/modules/logger/logged_topics.cpp
@@ -377,14 +377,14 @@ void LoggedTopics::add_high_rate_sensors_topics()
 	add_topic_multi("sensor_gps", 0, 4);
 	add_topic_multi("sensor_gnss_relative", 0, 1);
 
-	// add_topic("actuator_motors", 20);
-	// add_topic("vehicle_air_data", 20);
+	add_topic("actuator_motors", 20);
+	add_topic("vehicle_air_data", 20);
 	add_topic("estimator_aid_src_baro_hgt", 20);
 	add_topic("vehicle_magnetometer", 20);
 
-	add_topic("actuator_motors");
-	add_topic("esc_status");
-	add_topic("vehicle_air_data");
+	// add_topic("actuator_motors");
+	// add_topic("esc_status");
+	// add_topic("vehicle_air_data");
 }
 
 void LoggedTopics::add_mavlink_tunnel()
